Convert to a custom resource
